Final submitted
This is the first and probably only commit I will make for this
project. This is a VERY simple collection game I made in Java for a
project. I think the best part is the level creation method in the Room
class. Give it a look, it's based on Mojang's Catacomb Snatch source
code (google it, it's available for download I think)